backend/requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
SQLAlchemy==2.0.44
mysqlclient==2.2.7
beautifulsoup4==4.14.2
requests==2.32.5
pydantic==2.12.4
python-dotenv==1.2.1

# LangChain + Google Gemini
langchain-core==1.0.3
langchain-community==0.4.1
langchain-google-genai==3.0.1
langchain-text-splitters==1.0.0
